referral_system/pyproject.toml
[project]
name = "referral-system"
version = "0.1.0"
description = ""
authors = [
    {name = "elenaglu"}
]
readme = "README.md"
package-mode = false
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-phonenumber-field (>=8.0.0,<9.0.0)",
    "phonenumbers (>=9.0.0,<10.0.0)",
    "jwt (>=1.3.1,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "selenium (>=4.29.0,<5.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "allure-pytest (>=2.14.0,<3.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
